qpnp-charger: Enable vbatdet_lo IRQ on disconnect
This will cause the vbat_det lo IRQ to fire if pending.

It's safe to leave the IRQ enabled, since if it fires when
not charging, the IRQ handler will do nothing.  Also,
if you look at the code, this also will happen at EOC - the
IRQ will be enabled and never disabled until it fires.

Change-Id: I928e2db929f10d28dd37122b79bcc1853564e1c0