qpnp-charger: Fixup Oppo vbatdet configuration
Oppo sets vbatdet thresholds in the BQ24196 driver.  This isn't
working with the current code as that driver expects vbatdet as
an offset from vbatmax (either 100 or 300 mV).  This differs from
the standard qpnp-charger setup which wants a raw vbatdet threshold.

Most of these calls can just be removed in the future as the driver
never chooses anything other than 100 mV of offset, the exception
being when resume_charging is chosen - in which case the original
drivers seem to set a threshold of 100 mV above vddmax (which
effectively disables vbatdet...)

TODO:  Configure both the qpnp and BQ24196 vbatdet, since
the qpnp vbatdet is still used by some IRQs for driving
the charging state machine.

Change-Id: I84fb2cecbcc294345646190ab3fd53606e042fb0

Conflicts:
	drivers/power/qpnp-charger.c